rust: types: Implements PointerWrapper for primitives
For primitives that fit in a pointer size, kernel may use a pointer to
carry the value of these primitives, therefore make them PointerWrapper.

Note: the `borrow` API for PointerWrapper doesn't quite make sense for
primitives, so make it always return a reference to `()`, so that no one
will (ab)use it.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>