rust: add `ARef`, an owned reference to an always-refcounted object
This is meant to simplify how we define wrappers to types defined on the
C side of the source code. For example, `File` and `Credential`: we
currently have implementations of `Clone`, `Drop` for each plus a
"reference" type.

This patch simplifies this: types only need to implement a way to
increment and decrement their refcounts, then users can simply use `&T`
for shared references and `ARef<T>` for owned references. They can also
convert from one to the other (from `&T` to `ARef<T>` results in a
refcount increment).

Signed-off-by: Wedson Almeida Filho <wedsonaf@google.com>