rust: sync: Make {RwSemaphore, Mutex}::lock_noguard() safe
rw_semaphore and mutex set the owner task to their ->owner during the
lock aquisition, therefore to make `lock_noguard` safe, the ->owner must
not be freed if the lock is held.

Increases the refcount of ->owner in `lock_noguard` to avoid ->owner
getting freed, correspondingly, descreases the refcount in `unlock`.

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>