rust: add pin-init API
This API is used to facilitate pinned initialization of structs. Prior to this
API one had to use `unsafe` to create a partially initialized value and then
again use `unsafe` to initialize the value via pin-projections.
This API simplifies this process by combining these two steps into a macro
invocation with struct-initializer-like syntax.

Signed-off-by: Benno Lossin <y86-dev@protonmail.com>